Get rid of old DiceRoller class
