+++ src/main/java/dicerolling/DiceRoller.java
@@ -6,33 +6,23 @@ import logic.RandomColor;
 import org.javacord.api.entity.message.MessageAuthor;
 import org.javacord.api.entity.message.embed.EmbedBuilder;
 import sheets.PPManager;
-import statistics.PoolOptions;
-import statistics.RollResult;
+import statistics.RollResultBuilder;
 import statistics.resultvisitors.DifficultyVisitor;
 
-import java.util.ArrayList;
-import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
 import java.util.Random;
 import java.util.stream.Collectors;
 
 public class DiceRoller {
-
     public static final String NONE = "*none*";
-    private final PoolOptions poolOptions;
     private final Random random = new Random();
-    //The amount of doom generated by this dice roll
-    private final RollResult rollResult;
+    private final RollResultBuilder rollResult;
+    private final DicePool dicePool;
 
-    public DiceRoller(String content) {
-        //Split up content
-        ArrayList<String> args = new ArrayList<>(Arrays.asList(content.split(" ")));
-        args.remove("~r");
-        //Split dice into regular dice and plot dice
-        poolOptions = new PoolOptions();
-        DiceParameterHandler diceParameterHandler = new DiceParameterHandler(args, poolOptions);
-        diceParameterHandler.addDiceToPools();
-        rollResult = new RollResult(poolOptions.getTop());
+    public DiceRoller(DicePool dicePool) {
+        this.dicePool = dicePool;
+        rollResult = new RollResultBuilder(dicePool.getKeepHowMany(), false);
     }
 
     public int getDoom() {
@@ -46,19 +36,29 @@ public class DiceRoller {
         return buildResultEmbed(author, rollResult);
     }
 
-    private EmbedBuilder buildResultEmbed(MessageAuthor author, RollResult result) {
+    private EmbedBuilder buildResultEmbed(MessageAuthor author, RollResultBuilder result) {
         return new EmbedBuilder()
+                .setDescription("Here's the result for" + createDicePoolString(dicePool))
                 .setTitle(TwoDee.getRollTitleMessage())
                 .setAuthor(author)
                 .setColor(RandomColor.getRandomColor())
                 .addField("Regular dice", formatResults(result.getDice()), true)
-                .addField("Picked", resultsToString(result.getKept()), true)
+                .addField("Picked", resultsToString(result.getPickedDice()), true)
                 .addField("Dropped", resultsToString(result.getDropped()), true)
                 .addField("Plot dice", resultsToString(result.getPlotDice()), true)
                 .addField("Kept dice", resultsToString(result.getKeptDice()), true)
-                .addField("Flat bonuses", result.getFlatBonus() == 0 ? NONE : Integer.toString(result.getFlatBonus()), true)
-                .addField("Total", String.valueOf(result.getResult()), true)
-                .addField("Tier Hit", tiersHit(result.getResult()));
+                .addField("Flat bonuses", resultsToString(result.getFlatBonus()), true)
+                .addField("Total", String.valueOf(result.getTotal()), true)
+                .addField("Tier Hit", tiersHit(result.getTotal()));
+    }
+
+    private String createDicePoolString(DicePool dicePool) {
+        StringBuilder dicePoolString = new StringBuilder();
+        dicePool.getRegularDice().forEach(integer -> dicePoolString.append(" d").append(integer));
+        dicePool.getPlotDice().forEach(integer -> dicePoolString.append(" pd").append(integer));
+        dicePool.getKeptDice().forEach(integer -> dicePoolString.append(" kd").append(integer));
+        dicePool.getFlatBonuses().stream().map(integer -> (integer > 0) ? ("+" + integer) : integer).forEach(dicePoolString::append);
+        return dicePoolString.toString();
     }
 
     private String tiersHit(int total) {
@@ -86,7 +86,7 @@ public class DiceRoller {
     }
 
     //Bold 1s to show total doom generated. Runs doom increasing method afterwards.
-    private String formatResults(ArrayList<Integer> s) {
+    private String formatResults(List<Integer> s) {
         StringBuilder resultString = new StringBuilder();
         if (s.size() > 1) {
             for (int i = 0; i < s.size() - 1; i++) {
@@ -118,7 +118,12 @@ public class DiceRoller {
         return resultString.toString();
     }
 
-    // Username is stored as <@!140973544891744256>
+    /**
+     * Add 1 plot point if player generates an opportunity on their roll (at least one 1)
+     *
+     * @param author The player that rolled a 1
+     * @return The EmbedBuilder that shows the change in plot points for the player
+     */
     public EmbedBuilder addPlotPoints(MessageAuthor author) {
         if (rollResult.getDoom() != 0) {
             PPManager manager = new PPManager();
@@ -149,39 +154,30 @@ public class DiceRoller {
     }
 
     private void addFlatBonus() {
-        for (int flatBonus : poolOptions.getFlatBonus()) {
-            rollResult.addFlatBonuses(flatBonus);
-        }
+        dicePool.getFlatBonuses().forEach(rollResult::addFlatBonus);
     }
 
     private void rollPlotDice(Random random) {
         //A plot die's minimum value is its number of faces / 2
-        for (Integer pDice : poolOptions.getPlotDice()) {
-            int pValue = random.nextInt(pDice) + 1;
-            if (pValue < pDice / 2) {
-                pValue = pDice / 2;
+        for (Integer pdice : dicePool.getPlotDice()) {
+            int pValue = random.nextInt(pdice) + 1;
+            if (pValue < pdice / 2) {
+                pValue = pdice / 2;
             }
-            rollResult.addPlotDice(pValue);
+            rollResult.addPlotResult(pValue);
         }
     }
 
     private void rollDie(Random random) {
-        for (Integer die : poolOptions.getRegularDice()) {
-            int diceVal = random.nextInt(die) + 1;
-            rollResult.addDiceToResult(diceVal);
-        }
+        dicePool.getRegularDice().stream().mapToInt(die -> random.nextInt(die) + 1).forEach(rollResult::addResult);
     }
 
     private void rollKeptDie(Random random) {
-        for (Integer keptDie : poolOptions.getKeptDice()) {
-            int keptVal = random.nextInt(keptDie) + 1;
-            rollResult.addKeptDice(keptVal);
-        }
+        dicePool.getKeptDice().stream().mapToInt(keptDie -> random.nextInt(keptDie) + 1).forEach(rollResult::addKeptResult);
     }
 
     //Replaces brackets in the string. If the string is blank, returns "none" in italics
-    private String resultsToString(ArrayList<Integer> result) {
+    private String resultsToString(List<Integer> result) {
         return result.isEmpty() ? NONE : result.stream().map(Object::toString).collect(Collectors.joining(", "));
     }
-
 }
